게시글 삭제시 권한 확인

+++ src/main/java/com/example/prj1fe20231109mj/controller/BoardController.java
@@ -1,8 +1,11 @@
 package com.example.prj1fe20231109mj.controller;
 
 import com.example.prj1fe20231109mj.domain.Board;
+import com.example.prj1fe20231109mj.domain.Member;
 import com.example.prj1fe20231109mj.service.BoardService;
+import jakarta.servlet.http.HttpSession;
 import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
@@ -16,16 +19,22 @@ public class BoardController {
     private final BoardService service;
 
     @PostMapping("add")
-    public ResponseEntity add(@RequestBody Board board) {
+    public ResponseEntity add(@RequestBody Board board,
+                              @SessionAttribute(value = "login", required = false) Member login) {
+
+        System.out.println("login = " + login);
+
+        if (login == null) {
+
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+        }
 
         if (!service.validate(board))
             return ResponseEntity.badRequest().build();
-        if (service.save(board)) {
+        if (service.save(board, login)) {
             return ResponseEntity.ok().build();
         } else {
             return ResponseEntity.internalServerError().build();
-
-
         }
     }
 
@@ -39,8 +48,21 @@ public class BoardController {
         return service.get(id);
     }
 
+
+
+    // 게시물을 지울때!
     @DeleteMapping("remove/{id}")
-    public ResponseEntity remove(@PathVariable Integer id) {
+    public ResponseEntity remove(@PathVariable Integer id,
+                                 @SessionAttribute(value = "login", required = false) Member login) {
+
+        if (login == null) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build(); // 401
+        }
+
+        if (!service.hasAccess(id, login)){
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build(); // 403
+        }
+
         if (service.remove(id)) {
             return ResponseEntity.ok().build();
         } else {
